requirements.txt
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
execnet==1.9.0
Flask==2.0.2
glob2==0.7
h11==0.12.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
outcome==1.1.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
Pillow==8.4.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
pytest-bdd==5.0.0
pytest-forked==1.4.0
pytest-html-reporter==0.2.6
pytest-xdist==2.5.0
PyYAML==6.0
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
